requirements.txt
argparse
pycryptodome
pillow
numpy==1.19.3
prettytable
opencv-python
pandas
pyfiglet
image-quality
